Split layouts and move reusable compoments
