+++ frontend/app/routes/_protected+/_layout.tsx
@@ -1,6 +1,6 @@
 import { Outlet, isRouteErrorResponse, useRouteError } from '@remix-run/react';
 
-import ApplicationLayout, { NotFoundError, ServerError, i18nNamespaces as layoutI18nNamespaces } from '~/components/layouts/application-layout';
+import { NotFoundError, ProtectedLayout, ServerError, i18nNamespaces as layoutI18nNamespaces } from '~/components/layouts/protected-layout';
 import type { RouteHandleData } from '~/utils/route-utils';
 
 export const handle = {
@@ -13,18 +13,18 @@ export function ErrorBoundary() {
   if (isRouteErrorResponse(error)) {
     switch (error.status) {
       case 404: {
-        return <NotFoundError error={error} layout="protected" />;
+        return <NotFoundError error={error} />;
       }
     }
   }
 
-  return <ServerError error={error} layout="protected" />;
+  return <ServerError error={error} />;
 }
 
 export default function Layout() {
   return (
-    <ApplicationLayout layout="protected">
+    <ProtectedLayout>
       <Outlet />
-    </ApplicationLayout>
+    </ProtectedLayout>
   );
 }
